second-commit
Added checking of Id existing into Update/Delete functions, added DatabaseView into NavigationBar

+++ CitizenDevelopment1/Database/AccessToDatabase.cs
@@ -7,6 +7,7 @@ using System.Data;
 using System.Data.SQLite;
 using System.Threading.Tasks;
 using Dapper;
+using CitizenDevelopment1.Views;
 
 namespace CitizenDevelopment1.Database
 {
@@ -20,29 +21,93 @@ namespace CitizenDevelopment1.Database
             }
         }
 
-        public static void UpdateUser(int idConverted, User user)
+        public static bool UpdateUser(int idConverted, User user)
         {
             using (SQLiteConnection cnn = new SQLiteConnection(LoadConnectionString()))
             {
-                cnn.Open();
-                cnn.Execute($"update Users set applicationName=@applicationName, userName=@userName, comment=@comment where id={idConverted}", user);
-                cnn.Close();
+                bool isRecordExists = IsIdExists(idConverted);
+                if (isRecordExists)
+                {
+                    cnn.Open();
+                    cnn.Execute($"update Users set applicationName=@applicationName, userName=@userName, comment=@comment where id={idConverted}", user);
+                    cnn.Close();
+                    return true;
+                }
+                else
+                {
+                    return false;
+                }
             }
         }
 
-        public static void DeleteUser(int idConverted)
+        public static bool DeleteUser(int idConverted)
         {
             using (SQLiteConnection cnn = new SQLiteConnection(LoadConnectionString()))
             {
+                bool isRecordExists = IsIdExists(idConverted);
+                if (isRecordExists)
+                {
+                    cnn.Open();
+                    cnn.Execute($"delete from Users where id={idConverted}");
+                    cnn.Close();
+                    return true;
+                }
+                else
+                {
+                    return false;
+                }
+            }
+        }
+
+        public static List<UsersOutput> DatabaseUser()
+        {
+            using (SQLiteConnection cnn = new SQLiteConnection(LoadConnectionString()))
+            {
+                SQLiteCommand cmd = new SQLiteCommand();
+                List<UsersOutput> users = new List<UsersOutput>();
+
                 cnn.Open();
-                cnn.Execute($"delete from Users where id={idConverted}");
+
+                cmd.CommandText = "select * from Users";
+                cmd.Connection = cnn;
+
+                SQLiteDataReader dr;
+                dr = cmd.ExecuteReader();
+
+                while (dr.Read())
+                {
+                    users.Add(new UsersOutput(
+                        dr["id"].ToString(),
+                        dr["applicationName"].ToString(),
+                        dr["userName"].ToString(),
+                        dr["comment"].ToString()
+                        ));
+                 }
                 cnn.Close();
+                return users;
             }
         }
 
         private static string LoadConnectionString(string id = "DefaultConnection")
         {
             return ConfigurationManager.ConnectionStrings[id].ConnectionString;
+        }
+
+        public static bool IsIdExists(int idConverted)
+        {
+            using (SQLiteConnection cnn = new SQLiteConnection(LoadConnectionString()))
+            {
+                bool exists = false;
+                try
+                {
+                    exists = cnn.ExecuteScalar<bool>($"SELECT EXISTS(SELECT 1 FROM Users WHERE id={idConverted})");
+                }
+                catch (Exception)
+                {
+                    exists = false;
+                }
+                return exists;
+            }
         }
     }
 }
